[#44] 하나만 실행하게 의존성 관리 개선
