+++ Projects/Application/Monoree/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
