+++ Projects/Application/Monoree/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
