+++ Projects/Application/Monoree/Resources/Assets.xcassets/AppIcon.appiconset/search_icon.png
