+++ Projects/Application/Monoree/Resources/Assets.xcassets/Contents.json
