+++ Projects/Application/Monoree/Resources/Assets.xcassets/splash.imageset/Contents.json
