+++ Projects/Application/Monoree/Resources/Assets.xcassets/splash.imageset/splash.png
